refactor: remove AccessibilityControls tests and add full-featured and simple HTTP service examples
- Deleted the AccessibilityControls test file to streamline the codebase.
- Added a comprehensive full-featured service example demonstrating dual transport support (HTTP and gRPC) with advanced configurations and health checks.
- Introduced a simple HTTP service example showcasing basic HTTP functionality, routing, health checks, and environment configuration.
