+++ docs/pages/en/examples/full-featured.md
