+++ docs/pages/zh/examples/full-featured.md
